# colorbalancergb: optimization, improve vectorization