… src/common/dttypes.h
… src/iop/colorbalancergb.c