In an effort to speed up Color balance RGB, I took on a couple of low-hanging fixes / improvements to the pixel loop in `process()`, hoping it would allow modern compilers to vectorize the code. Currently Color balance RGB takes a significant portion of time compared to other modules typically present in a scene-referred workflow (on CPU) – 2.3 seconds compared to a total processing time of 3.6 seconds. GPU version is fine as-is.

```
$ darktable-cli images/mire1.cr2 0083-colorbalancergb/colorbalancergb.xmp output.png --core -d perf --disable-opencl
output file already exists, it will get renamed
1.724157 [dev] took 0.161 secs (0.139 CPU) to load the image.
1.851654 [export] creating pixelpipe took 0.124 secs (0.387 CPU)
1.851755 [dev_pixelpipe] took 0.000 secs (0.000 CPU) initing base buffer [export]
1.865977 [dev_pixelpipe] took 0.014 secs (0.032 CPU) processed `raw black/white point' on CPU, blended on CPU [export]
1.880293 [dev_pixelpipe] took 0.014 secs (0.027 CPU) processed `white balance' on CPU, blended on CPU [export]
1.885870 [dev_pixelpipe] took 0.006 secs (0.021 CPU) processed `highlight reconstruction' on CPU, blended on CPU [export]
2.027324 [dev_pixelpipe] took 0.141 secs (0.497 CPU) processed `demosaic' on CPU, blended on CPU [export]
2.062370 [dev_pixelpipe] took 0.035 secs (0.065 CPU) processed `exposure' on CPU, blended on CPU [export]
2.097796 [dev_pixelpipe] took 0.035 secs (0.134 CPU) processed `input color profile' on CPU, blended on CPU [export]
image colorspace transform Lab-->RGB took 0.046 secs (0.168 CPU) [channelmixerrgb ]
2.692354 [dev_pixelpipe] took 0.595 secs (2.314 CPU) processed `color calibration' on CPU, blended on CPU [export]
4.989018 [dev_pixelpipe] took 2.297 secs (9.015 CPU) processed `color balance rgb' on CPU, blended on CPU [export]
5.319613 [dev_pixelpipe] took 0.331 secs (1.271 CPU) processed `filmic rgb' on CPU, blended on CPU [export]
image colorspace transform RGB-->Lab took 0.036 secs (0.136 CPU) [colorout ]
5.427945 [dev_pixelpipe] took 0.108 secs (0.414 CPU) processed `output color profile' on CPU, blended on CPU [export]
5.489746 [dev_pixelpipe] took 0.062 secs (0.233 CPU) processed `display encoding' on CPU, blended on CPU [export]
5.489868 [dev_process_export] pixel pipeline processing took 3.638 secs (14.032 CPU)
```

Here's roughly what was done so far:

* Remove roundtrip `Yrg -> Ych -> test conversion to Yrg -> actual conversion to Yrg`. Instead perform hue rotation by linear algebra in `Yrg` and eliminate the test conversion in the gamut mapping phase. Usage of trigonometric functions is entirely eliminated in this phase of the calculations.
* Also eliminate explicit computation of sines and cosines elsewhere. This is nice since there is a potentially faster way to get the angles and also it should be easier to auto-vectorize (no need to call a vector math library)
* Remove branching from the gamut lookup routine (which is essentially a linear interpolation of a look-up table with wrapping at the edges). simplifying the control flow
* Replace `hypotf(x, y)` by `sqrtf(x*x + y*y)` – we don't need `hypotf()`'s handling of overflow here, and we avoid a slow library call – `sqrtf()` can probably be calculated using native instructions

These are all worthwhile changes alone, in my opinion, and are included here as separate commits. They make the code a bit faster, although nothing drastic (proper measurements to be done). However, as I was aiming at getting the whole pixel loop to an auto-vectorizable state, I also replaced the remaining calls to `atan2f()` by a custom branchless polynomial approximation. According to experiments in godbolt, neither GCC nor Clang seem to auto-vectorize `atan2f()` calls even if linking with a vector math library (`libmvec`, in a recent Clang version use `-fveclib=libmvec`).

With clang I'm indeed able to get the pixel loop auto-vectorized – well, almost. Two things that are preventing auto-vectorization are the lookup `opacities[g->mask_type]` in mask drawing, and the gamut LUT lookup. If I replace these with some dummy assignment, the loop auto-vectorizes and executes quite a bit faster than the non-vectorized version (a factor of 2-3). These dummy changes are included in the last commit along with notes. However, I couldn't find a way so far to convince Clang that there's no pointer aliasing taking place here (vectorization is failing with message `cannot identify array bounds`). @ralfbrown has done a bunch of nice work regarding auto-vectorization recently – would you like to have a look at this?

There's also the challenge of choosing a good optimization strategy. I'm not sure trying to get the whole loop to auto-vectorize is the best strategy – after all there are some operations in the loop that are naturally vectorized within one pixel: several `for_four_channels()` loops and `dot_product()`s at least. In other parts, individual channels of `Yrg` and `JzAzBz` are processed in a scalar-like manner and would benefit from vectorization, processing several pixels at a time. In principle the pixel loop could be split in several parts with different vectorization accordingly, but then one would have to allocate memory to store the `opacities[]` array that is calculated early in the processing.

Any help and comments are appreciated here. :)