@ralfbrown 2021-07-31
Often, you'll need to extract a line or two into a separate inline function to convince the compiler to vectorize.  For the particular case of `opacities[g->mask_type]`, I'd try something like
```
static inline void compensate_opacity(dt_aligned_pixel_t pix_out, const dt_aligned_pixel_t color, const float opacity)
{
   for_each_channel(c) pix_out[c] = (1.0f - opacity) * color[c] + opacity * MAX(pix_out[c], 0.0f);
}

...
  compensate_opacity(pix_out, color, opacities[g->mask_type]);
```
You'll note that I also changed the `fmaxf` to use the `MAX` macro instead.  I've typically found that GCC won't vectorize the `fmaxf`, though I have seen it do so -- I think it will if you give the correct compile flag so that special cases the function handles but not the CPU instruction become irrelevant [-ffinite-math-only or something like that]).

`sqrtf` does indeed vectorize to a single SSE instruction.  `powf` doesn't have a single CPU instruction, but GCC can be convinced to vectorize it.  If you look in src/develop/openmp_maths.h, you'll see a number of `omp declare simd` pragmas which are skipped for GCC; I've recently discovered that some of them (including powf) can indeed be used for GCC (a couple of the others will cause GCC to call the vector version but then fail to link because they don't actually exist in the library....)