@flannelhead 2021-07-31
Thanks for the pointers @ralfbrown ! I'm going to further experiment with this. What do you think of the idea of splitting the processing into several loops with alternating
```
#pragma omp for
```
and
```
#pragma omp for simd
```
depending on the contents?

> `powf` doesn't have a single CPU instruction, but GCC can be convinced to vectorize it.
Clang 12 can do this too if given the flag `-fveclib=libmvec`, see https://godbolt.org/z/qe4asjo1W. Perhaps something to consider overall when Darktable is compiled with a recent Clang.