@ralfbrown 2021-07-31
Keep in mind that there's considerable overhead in distributing tasks to the thread pool and then synchronizing on completion of the `for`, so it's best to keep the number of consecutive parallel for loops as small as practical.