@flannelhead 2021-08-01
So it seems to me that many parts of the current code that are meant to be vectorized, aren't vectorized properly at least by GCC.

According to experiments in godbolt, this seems to be mainly due to the `-fno-finite-math-only` flag that is used to build DT. This prevents GCC from vectorizing e.g. calls to `powf` and thus renders many of the intensive loops in Color balance RGB non-vectorizable. I have not yet found a way to revert that flag for a particular piece of code. `#pragma GCC optimize("finite-math-only")` doesn't seem effective.