@aurelienpierre 2021-08-02
> Remove roundtrip Yrg -> Ych -> test conversion to Yrg -> actual conversion to Yrg. Instead perform hue rotation by linear algebra in Yrg and eliminate the test conversion in the gamut mapping phase.

I don't understand and that sounds dangerous. How do you ensure the consistency of hue, then ? The test conversion is meant to compute find the max chroma available at current luminance and hue, if you stay in Yrg, r and g are both dependent of hue.