@flannelhead 2021-08-02
> I don't understand and that sounds dangerous. How do you ensure the consistency of hue, then ? The test conversion is meant to find the max chroma available at current luminance and hue, if you stay in Yrg, r and g are both dependent of hue.

My wording was maybe bad. The point was removing the explicit conversions requiring the computation of `atan2f`, `sinf`, `cosf` etc. which is expensive. Rather, do the hue rotation using a rotation matrix in `Yrg`, and compute the sin and cos of the hue angle from components r, g normalized by chroma. See here:

https://github.com/darktable-org/darktable/pull/9696/commits/f49e200e854bd5317758bfa08b4c0d69089a9938#diff-2759cbb0ba85e7ea7ff712187c53df30d5b48f624f36295e8909bd21d66247f9R566-R604

FWIW, the output is practically unaltered according to the integration test.

Anyway, this is a work in progress and there are several other things incoming, as soon as I get `gprof` working to identify the hotspots.