@aurelienpierre 2021-08-02
> FWIW, the output is practically unaltered according to the integration test.

I need to check what it is that the integration actually does, because I don't think there is something checking for out-of-gamut handling, which is the most important thing here.