@flannelhead 2021-08-02
Right. The out-of-gamut handling should be fine though, since I'm not removing it but replacing the mathematical expressions there with equivalent but computationally cheaper ones.