@flannelhead 2022-05-06
@aurelienpierre thanks for reminding. I think I'll close this one and proceed with optimizations in smaller chunks. In fact, there is already one: [converting from `dot_product()` to `dt_apply_transposed_color_matrix()`](https://github.com/darktable-org/darktable/pull/10963) but that alone is already quite big an undertaking since many code paths have to be updated.

Another one that could be worthwhile is eliminating Yrg <-> Ych conversions where applicable (as demonstrated in this PR) since that involves many calls to `sin()`, `cos()` and `atan2()` which are slow.

These two chunks I could probably manage to finish during May. Then there are smaller chunks here that can be considered but I'd prefer to land these one-by-one with measurements of the actual perf gains.